TicketingPage DateSelectUI Sticky 반영
