+++ src/Pages/Ticketing/TicketingPage.jsx
@@ -3,30 +3,36 @@ import InfoMain from '../../components/ui/performaceInfo/infoMain'
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '../../components/ui/tabs'
 import useEventInfo from '../../custom/useEventInfo'
 import DateSelectUI from '@/components/TicketingComponents/DateSelect'
+import Header from '@/components/Layout/Header'
+import Footer from '@/components/Layout/Footer'
 
 export default function TicketingPage() {
   const posterInfo = useEventInfo()
   return (
-    <div className='w-full flex flex-row items-center justify-center'>
-      <div className='w-2/3'>
-        <InfoMain />
-        <Tabs defaultValue='performaceInfo'>
-          <TabsList className=''>
-            <TabsTrigger value='performaceInfo'>공연정보</TabsTrigger>
-            <TabsTrigger value='review'>관람후기</TabsTrigger>
-            <TabsTrigger value='qna'>Q&A</TabsTrigger>
-          </TabsList>
-          <TabsContent value='performaceInfo'>
-            <PosterInfo src='./src/assets/jesusMainPic.jpg' />
-          </TabsContent>
-          <TabsContent value='review'> 관람후기 게시판 </TabsContent>
-          <TabsContent value='qna'> qna </TabsContent>
-        </Tabs>
+    <>
+      <Header />
+      <div className='w-full flex flex-row items-start justify-center mt-10'>
+        <div className='w-2/3'>
+          <InfoMain />
+          <Tabs defaultValue='performaceInfo'>
+            <TabsList className=''>
+              <TabsTrigger value='performaceInfo'>공연정보</TabsTrigger>
+              <TabsTrigger value='review'>관람후기</TabsTrigger>
+              <TabsTrigger value='qna'>Q&A</TabsTrigger>
+            </TabsList>
+            <TabsContent value='performaceInfo'>
+              <PosterInfo src='./src/assets/jesusMainPic.jpg' />
+            </TabsContent>
+            <TabsContent value='review'> 관람후기 게시판 </TabsContent>
+            <TabsContent value='qna'> qna </TabsContent>
+          </Tabs>
+        </div>
+        <div className='mt-10 ml-20 sticky top-48'>
+          <DateSelectUI />
+        </div>
       </div>
-      <div>
-        <DateSelectUI />
-      </div>
-    </div>
+      <Footer />
+    </>
   )
 }
