+++ src/components/Layout/Header.jsx
@@ -1,47 +1,11 @@
 import { Link } from 'react-router-dom'
 import { Input } from '@/components/ui/input'
 import { useState } from 'react'
-
-function SearchIcon(props) {
-  return (
-    <svg
-      {...props}
-      xmlns='http://www.w3.org/2000/svg'
-      width='24'
-      height='24'
-      viewBox='0 0 24 24'
-      fill='none'
-      stroke='currentColor'
-      strokeWidth='2'
-      strokeLinecap='round'
-      strokeLinejoin='round'
-    >
-      <circle cx='11' cy='11' r='8' />
-      <path d='m21 21-4.3-4.3' />
-    </svg>
-  )
-}
-
-const SearchZone = () => {
-  return (
-    <div className='items-start relative max-w-md'>
-      <Input
-        type='text'
-        placeholder='Search...'
-        className='rounded-full h-12 text-base border pl-5 pr-4 py-2 md:min-w-[450px]'
-      />
-      <div className='absolute inset-y-0 right-5 flex items-center'>
-        <SearchIcon className='w-5 h-5 text-muted-foreground' />
-      </div>
-    </div>
-  )
-}
-
-const LogoZone = () => {
-  return <div className='relative w-56 h-full bg-slate-400'>로고 영역(이미지 필요)</div>
-}
+import LogoZone from '@/components/Layout/LogoZone'
+import SearchZone from '@/components/Layout/SearchZone'
 
 const UserNavi = (isLogin) => {
+  isLogin = isLogin.isLogin
   return (
     <div className='flex flex-row justify-end items-center gap-5 h-4 rounded-md flex-1 text-nowrap'>
       {isLogin ? (
@@ -57,7 +21,7 @@ const UserNavi = (isLogin) => {
           </li>
 
           <li>
-            <Link to='/'>회원가입</Link>
+            <Link to='/register'>회원가입</Link>
           </li>
 
           <li>
@@ -122,12 +86,12 @@ const Header = () => {
   const [isLogin, setIsLogin] = useState(false)
 
   return (
-    <div className='flex flex-col items-center w-full min-w-[1152px] border-b border-solid'>
+    <div className='flex flex-col items-center w-full min-w-[1152px] border-b border-solid '>
       {/*  헤더 네비 */}
       <NaviZone>
         <LogoZone />
         <SearchZone />
-        <UserNavi isLogin={isLogin} />
+        <UserNavi isLogin={false} />
       </NaviZone>
 
       {/* 장르 영역 */}
