+++ src/index.css
@@ -136,9 +136,9 @@ button:focus-visible {
 }
 
 .login-page {
+  display: flex;
   justify-content: center;
   align-items: center;
-  margin-top: 20vh;
 }
 
 .seat-select-page {
@@ -153,5 +153,4 @@ button:focus-visible {
   display: flex;
   justify-content: center;
   align-items: center;
-  margin-top: 20vh;
 }
